+++ generators/dashboard/dashboard.go
@@ -1,13 +1,16 @@
 package dashboard
 
 import (
+	"embed"
 	"os/exec"
 	"text/template"
 
-	"github.com/gobuffalo/genny/v2"
-	"github.com/gobuffalo/genny/v2/gogen"
-	"github.com/gobuffalo/packr/v2"
+	"github.com/swiftcarrot/genny"
+	"github.com/swiftcarrot/genny/gogen"
 )
+
+//go:embed templates
+var templates embed.FS
 
 func New(opts *Options) (*genny.Generator, error) {
 	g := genny.New()
@@ -16,7 +19,8 @@ func New(opts *Options) (*genny.Generator, error) {
 		return g, err
 	}
 
-	if err := g.Box(packr.New("dashi:generators:dashboard", "../dashboard/templates")); err != nil {
+	err := g.Templates(&templates)
+	if err != nil {
 		return g, err
 	}
 
@@ -27,7 +31,7 @@ func New(opts *Options) (*genny.Generator, error) {
 	t := gogen.TemplateTransformer(data, helpers)
 	g.Transformer(t)
 
-	g.Transformer(genny.Dot())
+	g.Transformer(genny.Replace("$dot$", "."))
 	g.Command(exec.Command("yarn"))
 
 	return g, nil