+++ generators/new/new.go
@@ -1,15 +1,19 @@
 package new
 
 import (
+	"embed"
 	"html/template"
 	"os/exec"
 
-	"github.com/gobuffalo/genny/v2"
-	"github.com/gobuffalo/genny/v2/gogen"
-	"github.com/gobuffalo/packr/v2"
 	"github.com/swiftcarrot/dashi/generators/dashboard"
 	"github.com/swiftcarrot/dashi/generators/packages"
+	"github.com/swiftcarrot/dashi/generators/server"
+	"github.com/swiftcarrot/genny"
+	"github.com/swiftcarrot/genny/gogen"
 )
+
+//go:embed templates
+var templates embed.FS
 
 func New(opts *Options) (*genny.Group, error) {
 	err := opts.Validate()
@@ -20,12 +24,14 @@ func New(opts *Options) (*genny.Group, error) {
 	gg := &genny.Group{}
 
 	g := genny.New()
-	g.RunFn(func(r *genny.Runner) error {
-		r.Exec(exec.Command("go", "mod", "init"))
-		return nil
-	})
 
-	g.Box(packr.New("dashi:generators:new", "../new/templates"))
+	g.Command(exec.Command("go", "mod", "init"))
+	// TODO: lock package version
+	g.Command(exec.Command("go", "get", "github.com/swiftcarrot/dashi"))
+	g.Command(exec.Command("go", "get", "github.com/99designs/gqlgen@5ad012e3d7be1127706b9c8a3da0378df3a98ec1"))
+	// g.Command(exec.Command("go", "mod", "edit", "-replace=github.com/99designs/gqlgen=github.com/swiftcarrot/gqlgen@master"))
+
+	g.Templates(&templates)
 
 	data := map[string]interface{}{
 		"opts": opts,
@@ -34,10 +40,10 @@ func New(opts *Options) (*genny.Group, error) {
 	t := gogen.TemplateTransformer(data, helpers)
 	g.Transformer(t)
 
-	g.Transformer(genny.Dot())
+	g.Transformer(genny.Replace("$dot$", "."))
 	gg.Add(g)
 
-	if opts.APIOnly != true {
+	if !opts.APIOnly {
 		packages, err := packages.New(&packages.Options{
 			Name: opts.Name,
 		})
@@ -56,8 +62,18 @@ func New(opts *Options) (*genny.Group, error) {
 	}
 
 	make := genny.New()
-	make.Command(exec.Command("make"))
+	make.Command(exec.Command("make", "graphql"))
 	gg.Add(make)
+
+	server, err := server.New(&server.Options{
+		Name:    opts.Name,
+		Package: opts.Package,
+	})
+	if err != nil {
+		return nil, err
+	}
+	server.Command(exec.Command("go", "get"))
+	gg.Add(server)
 
 	return gg, nil
 }