+++ generators/scaffold/dashboard/dashboard.go
@@ -1,13 +1,17 @@
 package dashboard
 
 import (
+	"embed"
+	"os/exec"
 	"text/template"
 
-	"github.com/gobuffalo/genny/v2"
-	"github.com/gobuffalo/genny/v2/gogen"
-	"github.com/gobuffalo/packr/v2"
 	"github.com/swiftcarrot/dashi/generators/scaffold"
+	"github.com/swiftcarrot/genny"
+	"github.com/swiftcarrot/genny/gogen"
 )
+
+//go:embed templates/$pages$ templates/$components$
+var templates embed.FS
 
 func New(opts *scaffold.Options) (*genny.Generator, error) {
 	g := genny.New()
@@ -20,13 +24,14 @@ func New(opts *scaffold.Options) (*genny.Generator, error) {
 
 	name := opts.Name.Pascalize().Pluralize().ToLower().String()
 
-	g.Transformer(genny.Replace("-pages-", "packages/dashboard/src/pages/"+name))
-	g.Transformer(genny.Replace("-components-", "packages/components/src/"+name))
+	g.Transformer(genny.Replace("$pages$", "packages/dashboard/src/pages/"+name))
+	g.Transformer(genny.Replace("$components$", "packages/components/src/"+name))
 
-	err := g.Box(packr.New("scaffold:dashboard:templates", "../dashboard/templates"))
-	if err != nil {
+	if err := g.Templates(&templates); err != nil {
 		return g, err
 	}
+
+	g.Command(exec.Command("yarn", "fmt"))
 
 	return g, nil
 }