+++ generators/server/server.go
@@ -1,25 +1,32 @@
-package packages
+package server
 
 import (
 	"embed"
-	"html/template"
-	"os/exec"
+	"text/template"
 
+	"github.com/swiftcarrot/flect"
 	"github.com/swiftcarrot/genny"
 	"github.com/swiftcarrot/genny/gogen"
 )
 
-//go:embed templates
+type Options struct {
+	Name    flect.Ident
+	Package string
+}
+
+// Validate that options are usuable
+func (opts *Options) Validate() error {
+	return nil
+}
+
+//go:embed templates/api templates/cmd
 var templates embed.FS
 
 func New(opts *Options) (*genny.Generator, error) {
 	g := genny.New()
 
-	if err := opts.Validate(); err != nil {
-		return g, err
-	}
-
-	if err := g.Templates(&templates); err != nil {
+	err := g.Templates(&templates)
+	if err != nil {
 		return g, err
 	}
 
@@ -29,9 +36,6 @@ func New(opts *Options) (*genny.Generator, error) {
 	helpers := template.FuncMap{}
 	t := gogen.TemplateTransformer(data, helpers)
 	g.Transformer(t)
-
-	g.Transformer(genny.Replace("$dot$", "."))
-	g.Command(exec.Command("yarn"))
 
 	return g, nil
 }